Substract timestamps using ticks_diff() to account for wrap around (#36)
From the utime documentation, ticks_diff() should be used to compare
timestamps. They say the wrap around is implementation specific but on
forums I see people saying it's about 298 hours/12 days which at least
is the case for me on the tinypico.

In this specific case, the wrap around means the deltas will be small
negative numbers which in turn will make is_server_alive() return True
for days without sending anymore pings to the server. We will soon after
get disconnected and the application will pretty much stay forever in
that state.