Set _msg_id default value back to 0
This change got introduced in the python/mp split commit:
81dfc160a94c02ef03737cf00e426f69298dd148

Having it default to 1 is fine on the clear text port but doesn't
authenticate when connecting to the server through SSL.

The server will ack the token packet but won't reply anything.

Setting it back to 0 by default which mirrors what blynklib.py has.